[Android Viewer] Fix crash after save dialog
Fixed the crash after save dialog caused by
a possible null pointer related to presentation
mode.

+++ android/source/src/java/org/libreoffice/LibreOfficeMainActivity.java
@@ -412,7 +412,7 @@ public class LibreOfficeMainActivity extends AppCompatActivity implements Settin
                 // noinspection ResultOfMethodCallIgnored
                 mTempFile.delete();
             }
-            if (mTempSlideShowFile.exists()) {
+            if (mTempSlideShowFile != null && mTempSlideShowFile.exists()) {
                 // noinspection ResultOfMethodCallIgnored
                 mTempSlideShowFile.delete();
             }
